Updating blender file
