+++ .gitignore
@@ -102,3 +102,4 @@ dist
 
 # TernJS port file
 .tern-port
+*.blend1
